Move SingleComponentModel to a new module
