+++ geometry_constructor/qml_models/component_filters.py
@@ -1,0 +1,35 @@
+from PySide2.QtCore import Property, QModelIndex, QSortFilterProxyModel, Signal
+
+
+class SingleComponentModel(QSortFilterProxyModel):
+    """A filtered model that only displays a single component from an InstrumentModel"""
+
+    def __init__(self):
+        super().__init__()
+        self.desired_index = 0
+
+    def get_index(self):
+        return self.desired_index
+
+    def set_index(self, val):
+        self.desired_index = val
+        self.invalidateFilter()
+
+    index_changed = Signal()
+
+    index = Property(int, get_index, set_index, notify=index_changed)
+
+    def get_model(self):
+        return self.sourceModel()
+
+    def set_model(self, val):
+        self.setSourceModel(val)
+        self.invalidateFilter()
+
+    model_changed = Signal()
+
+    model = Property('QVariant', get_model, set_model, notify=model_changed)
+
+    def filterAcceptsRow(self, source_row: int, source_parent: QModelIndex):
+        """Overrides filterAcceptsRow to only accept the component at the given index"""
+        return source_row == self.desired_index
