+++ geometry_constructor/qml_models/instrument_model.py
@@ -3,7 +3,7 @@ from geometry_constructor.data_model import ComponentType, PixelGrid, PixelMappi
 from geometry_constructor.qml_models import change_value, generate_unique_name
 from geometry_constructor.qml_models.transform_model import TransformationModel
 from geometry_constructor.off_renderer import OffMesh
-from PySide2.QtCore import Property, Qt, QAbstractListModel, QModelIndex, QSortFilterProxyModel, Signal, Slot
+from PySide2.QtCore import Qt, QAbstractListModel, QModelIndex, Signal, Slot
 from PySide2.QtGui import QMatrix4x4, QVector3D
 
 
@@ -351,37 +351,3 @@ class InstrumentModel(QAbstractListModel):
         """
         model_index = self.createIndex(index, 0)
         self.dataChanged.emit(model_index, model_index, InstrumentModel.MeshRole)
-
-
-class SingleComponentModel(QSortFilterProxyModel):
-    """A filtered model that only displays a single component from an InstrumentModel"""
-
-    def __init__(self):
-        super().__init__()
-        self.desired_index = 0
-
-    def get_index(self):
-        return self.desired_index
-
-    def set_index(self, val):
-        self.desired_index = val
-        self.invalidateFilter()
-
-    index_changed = Signal()
-
-    index = Property(int, get_index, set_index, notify=index_changed)
-
-    def get_model(self):
-        return self.sourceModel()
-
-    def set_model(self, val):
-        self.setSourceModel(val)
-        self.invalidateFilter()
-
-    model_changed = Signal()
-
-    model = Property('QVariant', get_model, set_model, notify=model_changed)
-
-    def filterAcceptsRow(self, source_row: int, source_parent: QModelIndex):
-        """Overrides filterAcceptsRow to only accept the component at the given index"""
-        return source_row == self.desired_index
